baroque/pyproject.toml
[project]
name = "baroque"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=2.2",
    "pyyaml>=6.0",
]
